+++ CHANGELOG.md
@@ -4,6 +4,13 @@ This file documents all notable changes to this project. The format of this file
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- A new endpoint `GET /api/template` to fetch template(s) from template table.
+- A new endpoint `POST /api/template` to create a new template.
+
 ## [0.7.0] - 2020-01-28
 
 ### Added
@@ -263,6 +270,7 @@ v0.4.1
   - `ETCD_ADDR`
   - `ETCD_SIG_KEY`
 
+[Unreleased]: https://github.com/petabi/review/compare/0.7.0...master
 [0.7.0]: https://github.com/petabi/review/compare/0.6.3...0.7.0
 [0.6.3]: https://github.com/petabi/review/compare/0.6.2...0.6.3
 [0.6.2]: https://github.com/petabi/review/compare/0.6.1...0.6.2